requirements.txt
fastapi
uvicorn
sentence-transformers
qdrant-client
httpx
sqlalchemy
pydantic
python-dotenv
newsapi-python
yfinance
sqlalchemy>=2.0
